migrator: extract non-cli specifics from cli package
