+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -255,8 +255,11 @@ go_library(
         "//internal/conf/deploy",
         "//internal/conf/reposource",
         "//internal/database",
+        "//internal/database/migration",
         "//internal/database/migration/cliutil",
         "//internal/database/migration/drift",
+        "//internal/database/migration/multiversion",
+        "//internal/database/migration/runner",
         "//internal/database/migration/schemas",
         "//internal/deviceid",
         "//internal/encryption",
