+++ cmd/migrator/shared/BUILD.bazel
@@ -2,31 +2,19 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "shared",
-    srcs = [
-        "conf.go",
-        "main.go",
-        "registration.go",
-    ],
+    srcs = ["main.go"],
     importpath = "github.com/sourcegraph/sourcegraph/cmd/migrator/shared",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/conf/conftypes",
-        "//internal/database",
-        "//internal/database/basestore",
-        "//internal/database/connections/live",
+        "//internal/database/migration",
         "//internal/database/migration/cliutil",
+        "//internal/database/migration/runner",
         "//internal/database/migration/schemas",
         "//internal/database/migration/store",
-        "//internal/database/postgresdsn",
-        "//internal/jsonc",
         "//internal/observation",
-        "//internal/oobmigration",
         "//internal/oobmigration/migrations",
         "//internal/version",
-        "//lib/errors",
         "//lib/output",
-        "//schema",
-        "@com_github_keegancsmith_sqlf//:sqlf",
         "@com_github_sourcegraph_log//:log",
         "@com_github_urfave_cli_v2//:cli",
     ],
