+++ enterprise/cmd/frontend/BUILD.bazel
@@ -16,6 +16,7 @@ go_library(
     deps = [
         "//enterprise/cmd/frontend/shared",
         "//enterprise/cmd/sourcegraph/enterprisecmd",
+        "//internal/oobmigration",
         "//internal/sanitycheck",
         "//ui/assets",
         "//ui/assets/enterprise",
