+++ enterprise/cmd/sourcegraph/enterprisecmd/enterprisecmd.go
@@ -4,7 +4,6 @@
 package enterprisecmd
 
 import (
-	"github.com/sourcegraph/sourcegraph/internal/oobmigration"
 	"github.com/sourcegraph/sourcegraph/internal/service"
 	"github.com/sourcegraph/sourcegraph/internal/service/svcmain"
 )
@@ -20,9 +19,4 @@ func MainEnterprise(services []service.Service, args []string) {
 // enterprise (non-OSS) build to start a single service (such as frontend or gitserver).
 func SingleServiceMainEnterprise(service service.Service) {
 	svcmain.SingleServiceMain(service, config)
-}
-
-func init() {
-	// TODO(sqs): TODO(single-binary): could we move this out of init?
-	oobmigration.ReturnEnterpriseMigrations = true
 }
