+++ internal/database/migration/multiversion/BUILD.bazel
@@ -1,44 +1,28 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "cliutil",
+    name = "multiversion",
     srcs = [
-        "addlog.go",
-        "describe.go",
-        "describe_util.go",
-        "downgrade.go",
-        "downto.go",
         "drift.go",
-        "drift_autofix.go",
-        "help.go",
-        "iface.go",
-        "multiversion.go",
-        "run_oobmigrations.go",
-        "undo.go",
-        "up.go",
-        "upgrade.go",
-        "upto.go",
+        "plan.go",
+        "run.go",
         "util.go",
-        "validate.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/database/migration/cliutil",
+    importpath = "github.com/sourcegraph/sourcegraph/internal/database/migration/multiversion",
     visibility = ["//:__subpackages__"],
     deps = [
         "//internal/database",
         "//internal/database/migration/definition",
         "//internal/database/migration/drift",
-        "//internal/database/migration/multiversion",
         "//internal/database/migration/runner",
         "//internal/database/migration/schemas",
+        "//internal/database/migration/shared",
         "//internal/database/migration/store",
         "//internal/observation",
         "//internal/oobmigration",
         "//internal/oobmigration/migrations",
-        "//internal/version",
         "//internal/version/upgradestore",
         "//lib/errors",
         "//lib/output",
-        "@com_github_sourcegraph_log//:log",
-        "@com_github_urfave_cli_v2//:cli",
     ],
 )
