+++ internal/database/migration/store/BUILD.bazel
@@ -6,24 +6,35 @@ go_library(
     srcs = [
         "describe.go",
         "describe_scan.go",
+        "extractor.go",
         "observability.go",
+        "registration.go",
         "store.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/internal/database/migration/store",
     visibility = ["//:__subpackages__"],
     deps = [
+        "//internal/conf/conftypes",
+        "//internal/database",
         "//internal/database/basestore",
         "//internal/database/dbutil",
         "//internal/database/locker",
         "//internal/database/migration/definition",
+        "//internal/database/migration/runner",
         "//internal/database/migration/schemas",
         "//internal/database/migration/shared",
+        "//internal/database/postgresdsn",
+        "//internal/jsonc",
         "//internal/metrics",
         "//internal/observation",
+        "//internal/oobmigration",
+        "//internal/oobmigration/migrations",
         "//lib/errors",
+        "//schema",
         "@com_github_jackc_pgconn//:pgconn",
         "@com_github_keegancsmith_sqlf//:sqlf",
         "@com_github_lib_pq//:pq",
+        "@com_github_sourcegraph_log//:log",
         "@io_opentelemetry_go_otel//attribute",
     ],
 )
