+++ internal/database/migration/store/registration.go
@@ -1,4 +1,4 @@
-package shared
+package store
 
 import (
 	"context"
@@ -11,9 +11,44 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database/migration/schemas"
 	"github.com/sourcegraph/sourcegraph/internal/database/postgresdsn"
 	"github.com/sourcegraph/sourcegraph/internal/jsonc"
+	"github.com/sourcegraph/sourcegraph/internal/oobmigration"
+	"github.com/sourcegraph/sourcegraph/internal/oobmigration/migrations"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 	"github.com/sourcegraph/sourcegraph/schema"
 )
+
+type RegisterMigratorsUsingConfAndStoreFactoryFunc func(
+	ctx context.Context,
+	db database.DB,
+	runner *oobmigration.Runner,
+	conf conftypes.UnifiedQuerier,
+	storeFactory migrations.StoreFactory,
+) error
+
+func ComposeRegisterMigratorsFuncs(fnsFromConfAndStoreFactory ...RegisterMigratorsUsingConfAndStoreFactoryFunc) func(storeFactory migrations.StoreFactory) oobmigration.RegisterMigratorsFunc {
+	return func(storeFactory migrations.StoreFactory) oobmigration.RegisterMigratorsFunc {
+		return func(ctx context.Context, db database.DB, runner *oobmigration.Runner) error {
+			conf, err := newStaticConf(ctx, db)
+			if err != nil {
+				return err
+			}
+
+			fns := make([]oobmigration.RegisterMigratorsFunc, 0, len(fnsFromConfAndStoreFactory))
+			for _, f := range fnsFromConfAndStoreFactory {
+				f := f // avoid loop capture
+				if f == nil {
+					continue
+				}
+
+				fns = append(fns, func(ctx context.Context, db database.DB, runner *oobmigration.Runner) error {
+					return f(ctx, db, runner, conf, storeFactory)
+				})
+			}
+
+			return oobmigration.ComposeRegisterMigratorsFuncs(fns...)(ctx, db, runner)
+		}
+	}
+}
 
 type staticConf struct {
 	serviceConnections conftypes.ServiceConnections
@@ -21,7 +56,7 @@ type staticConf struct {
 }
 
 func newStaticConf(ctx context.Context, db database.DB) (*staticConf, error) {
-	serviceConnections, err := serviceConnections()
+	serviceConnections, err := migrationServiceConnections()
 	if err != nil {
 		return nil, err
 	}
@@ -40,7 +75,7 @@ func newStaticConf(ctx context.Context, db database.DB) (*staticConf, error) {
 func (c staticConf) ServiceConnections() conftypes.ServiceConnections { return c.serviceConnections }
 func (c staticConf) SiteConfig() schema.SiteConfiguration             { return c.siteConfig }
 
-func serviceConnections() (conftypes.ServiceConnections, error) {
+func migrationServiceConnections() (conftypes.ServiceConnections, error) {
 	dsns, err := postgresdsn.DSNsBySchema(schemas.SchemaNames)
 	if err != nil {
 		return conftypes.ServiceConnections{}, err
