+++ internal/oobmigration/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "interrupts.go",
         "migrator.go",
         "observability.go",
+        "progress.go",
         "registration.go",
         "runner.go",
         "store.go",
@@ -31,6 +32,7 @@ go_library(
         "//internal/version",
         "//internal/version/upgradestore",
         "//lib/errors",
+        "//lib/output",
         "@com_github_derision_test_glock//:glock",
         "@com_github_jackc_pgconn//:pgconn",
         "@com_github_keegancsmith_sqlf//:sqlf",
