+++ internal/version/upgradestore/store_test.go
@@ -22,7 +22,7 @@ func TestGetServiceVersion(t *testing.T) {
 	store := New(db)
 
 	t.Run("fresh db", func(t *testing.T) {
-		_, ok, err := store.GetServiceVersion(ctx, "service")
+		_, ok, err := store.GetServiceVersion(ctx)
 		if err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
@@ -32,17 +32,17 @@ func TestGetServiceVersion(t *testing.T) {
 	})
 
 	t.Run("after updates", func(t *testing.T) {
-		if err := store.UpdateServiceVersion(ctx, "service", "1.2.3"); err != nil {
+		if err := store.UpdateServiceVersion(ctx, "1.2.3"); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
-		if err := store.UpdateServiceVersion(ctx, "service", "1.2.4"); err != nil {
+		if err := store.UpdateServiceVersion(ctx, "1.2.4"); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
-		if err := store.UpdateServiceVersion(ctx, "service", "1.3.0"); err != nil {
+		if err := store.UpdateServiceVersion(ctx, "1.3.0"); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
 
-		version, ok, err := store.GetServiceVersion(ctx, "service")
+		version, ok, err := store.GetServiceVersion(ctx)
 		if err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
@@ -59,7 +59,7 @@ func TestGetServiceVersion(t *testing.T) {
 			t.Fatalf("unexpected error: %s", err)
 		}
 
-		_, ok, err := store.GetServiceVersion(ctx, "service")
+		_, ok, err := store.GetServiceVersion(ctx)
 		if err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
@@ -75,15 +75,15 @@ func TestSetServiceVersion(t *testing.T) {
 	db := database.NewDB(logger, dbtest.NewDB(logger, t))
 	store := New(db)
 
-	if err := store.UpdateServiceVersion(ctx, "service", "1.2.3"); err != nil {
+	if err := store.UpdateServiceVersion(ctx, "1.2.3"); err != nil {
 		t.Fatalf("unexpected error: %s", err)
 	}
 
-	if err := store.SetServiceVersion(ctx, "service", "1.2.5"); err != nil {
+	if err := store.SetServiceVersion(ctx, "1.2.5"); err != nil {
 		t.Fatalf("unexpected error: %s", err)
 	}
 
-	version, _, err := store.GetServiceVersion(ctx, "service")
+	version, _, err := store.GetServiceVersion(ctx)
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
 	}
@@ -99,7 +99,7 @@ func TestGetFirstServiceVersion(t *testing.T) {
 	store := New(db)
 
 	t.Run("fresh db", func(t *testing.T) {
-		_, ok, err := store.GetFirstServiceVersion(ctx, "service")
+		_, ok, err := store.GetFirstServiceVersion(ctx)
 		if err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
@@ -109,17 +109,17 @@ func TestGetFirstServiceVersion(t *testing.T) {
 	})
 
 	t.Run("after updates", func(t *testing.T) {
-		if err := store.UpdateServiceVersion(ctx, "service", "1.2.3"); err != nil {
+		if err := store.UpdateServiceVersion(ctx, "1.2.3"); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
-		if err := store.UpdateServiceVersion(ctx, "service", "1.2.4"); err != nil {
+		if err := store.UpdateServiceVersion(ctx, "1.2.4"); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
-		if err := store.UpdateServiceVersion(ctx, "service", "1.3.0"); err != nil {
+		if err := store.UpdateServiceVersion(ctx, "1.3.0"); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
 
-		firstVersion, ok, err := store.GetFirstServiceVersion(ctx, "service")
+		firstVersion, ok, err := store.GetFirstServiceVersion(ctx)
 		if err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
@@ -136,7 +136,7 @@ func TestGetFirstServiceVersion(t *testing.T) {
 			t.Fatalf("unexpected error: %s", err)
 		}
 
-		_, ok, err := store.GetFirstServiceVersion(ctx, "service")
+		_, ok, err := store.GetFirstServiceVersion(ctx)
 		if err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
@@ -163,12 +163,12 @@ func TestUpdateServiceVersion(t *testing.T) {
 			{"0.2.0", nil},
 			{"1.0.0", nil},
 			{"1.2.0", &UpgradeError{
-				Service:  "service",
+				Service:  "frontend",
 				Previous: semver.MustParse("1.0.0"),
 				Latest:   semver.MustParse("1.2.0"),
 			}},
 			{"2.1.0", &UpgradeError{
-				Service:  "service",
+				Service:  "frontend",
 				Previous: semver.MustParse("1.0.0"),
 				Latest:   semver.MustParse("2.1.0"),
 			}},
@@ -176,12 +176,12 @@ func TestUpdateServiceVersion(t *testing.T) {
 			{"non-semantic-version-is-always-valid", nil},
 			{"1.0.0", nil}, // back to semantic version is allowed
 			{"2.1.0", &UpgradeError{
-				Service:  "service",
+				Service:  "frontend",
 				Previous: semver.MustParse("1.0.0"),
 				Latest:   semver.MustParse("2.1.0"),
 			}}, // upgrade policy violation returns
 		} {
-			have := store.UpdateServiceVersion(ctx, "service", tc.version)
+			have := store.UpdateServiceVersion(ctx, tc.version)
 			want := tc.err
 
 			if !errors.Is(have, want) {
@@ -197,7 +197,7 @@ func TestUpdateServiceVersion(t *testing.T) {
 			t.Fatalf("unexpected error: %s", err)
 		}
 
-		if err := store.UpdateServiceVersion(ctx, "service", "0.0.1"); err == nil {
+		if err := store.UpdateServiceVersion(ctx, "0.0.1"); err == nil {
 			t.Fatalf("expected error, got none")
 		}
 	})
@@ -214,7 +214,7 @@ func TestValidateUpgrade(t *testing.T) {
 			t.Fatalf("unexpected error: %s", err)
 		}
 
-		if err := store.ValidateUpgrade(ctx, "service", "0.0.1"); err != nil {
+		if err := store.ValidateUpgrade(ctx, "frontend", "0.0.1"); err != nil {
 			t.Fatalf("unexpected error: %s", err)
 		}
 	})
