+++ lib/output/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "capabilities.go",
         "emoji.go",
         "line.go",
+        "logger.go",
         "noop_writer.go",
         "output.go",
         "output_unix.go",
@@ -37,6 +38,7 @@ go_library(
         "@com_github_mattn_go_runewidth//:go-runewidth",
         "@com_github_moby_term//:term",
         "@com_github_muesli_termenv//:termenv",
+        "@com_github_sourcegraph_log//:log",
         "@org_golang_x_term//:term",
     ] + select({
         "@io_bazel_rules_go//go/platform:windows": [
