+++ lib/output/emoji.go
@@ -1,5 +1,20 @@
 package output
 
+var allEmojis = [...]string{
+	EmojiFailure,
+	EmojiWarning,
+	EmojiSuccess,
+	EmojiInfo,
+	EmojiLightbulb,
+	EmojiAsterisk,
+	EmojiWarningSign,
+	EmojiFingerPointRight,
+	EmojiHourglass,
+	EmojiShrug,
+	EmojiOk,
+	EmojiQuestionMark,
+}
+
 // Standard emoji for use in output.
 const (
 	EmojiFailure          = "❌"
